# docs/swagger.yaml
openapi: 3.0.0
info:
  title: FoodRecommend API
  description: API Documentation for Food Decision Platform
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /api/health:
    get:
      summary: Health Check
      responses:
        '200':
          description: OK
  /api/auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                fullName:
                  type: string
      responses:
        '200':
          description: Created
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      token:
                        type: string
  /api/loyalty/info:
    get:
      summary: Get user loyalty info
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
  /api/loyalty/referral:
    post:
      summary: Apply referral code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Success
  /api/payment/zalopay/create:
    post:
      summary: Create ZaloPay payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                groupOrderId:
                  type: string
      responses:
        '200':
          description: Success

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT